telegram-reminder-bot-example/requirements.txt
anyio==4.1.0
APScheduler==3.10.4
certifi==2023.11.17
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
python-telegram-bot==20.7
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
tzlocal==5.2
